remove OptionalDrawStage
