+++ src/main/java/com/darzalgames/libgdxtools/maingame/StageLikeRenderable.java
@@ -12,7 +12,6 @@ public interface StageLikeRenderable extends InputProcessor {
 
 	void clear();
 
-	void setShouldDraw(boolean shouldDraw);
 	void resize(int width, int height);
 	Vector2 screenToStageCoordinates(Vector2 vector2);
